requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Faker==12.0.1
idna==3.3
importlib-metadata==4.11.1
iniconfig==1.1.1
Markdown==3.3.6
mixer==7.2.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.1.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
responses==0.20.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2021.5
urllib3==1.26.9
zipp==3.7.0
